1448501: subman gui can unregister, when network is up again
* When subman gui is started without working network, then
  file /etc/resolv.conf does not include any record for
  DNS server. When network is up again, then glibc does not
  reread this file automaticaly and function getaddrinfo()
  fails. It is necessary to reread this file explicitly
  using reset_resolver()
* I tried to triger reset_resolver() using D-Bus signals
  from NetworkManager, but it didn't work in subman-gui
  for some reason. Simple example of this approach is in
  this gist:

https://gist.github.com/jirihnidek/4759643a12abcd5d39f41d9d206512d5

* More information about reseting resolv.conf:

https://stackoverflow.com/questions/30360029/if-getaddrinfo-fails-once-it-fails-forever-even-after-network-is-ready
